+++ Saempl/Source/SampleFileFilterRuleLoudnessDecibel.h
@@ -30,13 +30,13 @@ public:
     /**
      @returns the compare value of the rule.
      */
-    int getCompareValue();
+    double getCompareValue();
     /**
      Sets the compare value of the rule.
      
      @param inCompareValue the value to set.
      */
-    void setCompareValue(int const & inCompareValue);
+    void setCompareValue(double const & inCompareValue);
     
 private:
     double mCompareValue;
